+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to `homebridge-config-ui-x` will be documented in this file.
 - added screen for ui v5 readiness check
 - add color temperature to lightbulb screen (@justjam2013)
 - add temperature units to heatercooler (@justjam2013)
+- add valve control modal (@justjam2013)
 
 ### Homebridge Dependencies